+++ src/Makefile.mingw
@@ -31,7 +31,7 @@ WINDRES= i686-w64-mingw32-windres
 TARGET = openmrac.exe
 RES    = omicon.res
 # echo -n "OBJS   = "; ls *.cpp | sed 's/.cpp$/.o/g' | tr '\n' ' ' ; echo
-OBJS   = 3dm.o cam.o car2d.o collider.o controls.o cstring1.o datkey.o fopendir.o gameaux.o gamemenu.o gamemng.o gamemng_frame.o gamemng_load.o gbuff_in.o ghost.o glext1.o glhelpers1.o gltext.o gridlayout.o gui.o guifont.o load_texture.o main.o mainmenu.o matmng.o minial.o octopus.o particles.o pict2.o pict2_jpeg.o pict2_png.o rand1.o rbsolver.o settings_dat.o settingsdialog.o skysph.o soundmng.o triboxint.o 
+OBJS   = 3dm.o cam.o car2d.o collider.o controls.o cstring1.o datkey.o dpiaware_win32.o fopendir.o gameaux.o gamemenu.o gamemng.o gamemng_frame.o gamemng_load.o gbuff_in.o ghost.o glext1.o glhelpers1.o gltext.o gridlayout.o gui.o guifont.o load_texture.o main.o mainmenu.o matmng.o minial.o octopus.o particles.o pict2.o pict2_jpeg.o pict2_png.o rand1.o rbsolver.o settings_dat.o settingsdialog.o skysph.o soundmng.o triboxint.o
 
 .PHONY: all clean install uninstall deb
