+++ src/dpiaware_win32.cpp
@@ -1,0 +1,49 @@
+// Disable DPI scaling on Windows
+
+#if defined(__WIN32__)
+
+#include <windows.h>
+
+#ifndef DPI_ENUMS_DECLARED
+typedef enum PROCESS_DPI_AWARENESS
+{
+    PROCESS_DPI_UNAWARE = 0,
+    PROCESS_SYSTEM_DPI_AWARE = 1,
+    PROCESS_PER_MONITOR_DPI_AWARE = 2
+} PROCESS_DPI_AWARENESS;
+#endif
+
+typedef BOOL (WINAPI * SETPROCESSDPIAWARE_T)(void);
+typedef HRESULT (WINAPI * SETPROCESSDPIAWARENESS_T)(PROCESS_DPI_AWARENESS);
+
+bool win32_SetProcessDpiAware(void) {
+    HMODULE shcore = LoadLibraryA("Shcore.dll");
+    SETPROCESSDPIAWARENESS_T SetProcessDpiAwareness = NULL;
+    if (shcore) {
+        SetProcessDpiAwareness = (SETPROCESSDPIAWARENESS_T) GetProcAddress(shcore, "SetProcessDpiAwareness");
+    }
+    HMODULE user32 = LoadLibraryA("User32.dll");
+    SETPROCESSDPIAWARE_T SetProcessDPIAware = NULL;
+    if (user32) {
+        SetProcessDPIAware = (SETPROCESSDPIAWARE_T) GetProcAddress(user32, "SetProcessDPIAware");
+    }
+
+    bool ret = false;
+    if (SetProcessDpiAwareness) {
+        ret = SetProcessDpiAwareness(PROCESS_PER_MONITOR_DPI_AWARE) == S_OK;
+    } else if (SetProcessDPIAware) {
+        ret = SetProcessDPIAware() != 0;
+    }
+
+    if (user32) {
+        FreeLibrary(user32);
+    }
+    if (shcore) {
+        FreeLibrary(shcore);
+    }
+    return ret;
+}
+
+volatile bool ha = win32_SetProcessDpiAware();
+
+#endif
